Pass cokies to error sites

+++ app.js
@@ -44,11 +44,11 @@ app.post('/login', async (req, res) => {
         res.redirect('/');
       }
     } else {
-      res.render('login', { message: 'Incorrect username or password' });
+      res.render('login', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Incorrect username or password' });
     }
   } catch (error) {
     console.error('Error during login:', error);
-    res.render('error', { message: 'An error occurred during login. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'An error occurred during login. Please try again.' });
   }
 });
 
@@ -76,7 +76,7 @@ app.post('/sign_up', async (req, res) => {
   }
   catch (err) {
     console.error('Error during sign up:', err);
-    res.render('error', { message: "Unable to add new user" });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: "Unable to add new user" });
   }
   res.redirect('/login');
 });
@@ -88,7 +88,7 @@ app.get('/cart', authorize.authorize_user, async (req, res) => {
     res.render('user/cart', { products: products, user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role });
   } catch (err) {
     console.error('Error showing cart:', err);
-    res.render('error', { message: 'Unable to show cart' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to show cart' });
   }
 });
 
@@ -99,7 +99,7 @@ app.get('/cart/clear', authorize.authorize_user, async (req, res) => {
     res.redirect('/cart');
   } catch (err) {
     console.error('Error clearing cart:', err);
-    res.render('error', { message: 'Unable to clear cart. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to clear cart. Please try again.' });
   }
 });
 
@@ -124,7 +124,7 @@ app.get('/cart/add/:id', authorize.authorize_user, async (req, res) => {
     }
   } catch (err) {
     console.error('Error adding to cart:', err);
-    res.render('error', { message: 'Unable to add to cart. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to add to cart. Please try again.' });
   }
 });
 
@@ -136,7 +136,7 @@ app.get('/cart/delete/:id', authorize.authorize_user, async (req, res) => {
     res.redirect('/cart');
   } catch (err) {
     console.error('Error deleting from cart:', err);
-    res.render('error', { message: 'Unable to delete from cart. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to delete from cart. Please try again.' });
   }
 });
 
@@ -149,7 +149,7 @@ app.post('/cart/submit', authorize.authorize_user, async (req, res) => {
     res.render('user/cart', { products: {}, user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role });
   } catch (err) {
     console.error('Error submitting cart:', err);
-    res.render('error', { message: 'Unable to submit cart. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to submit cart. Please try again.' });
   }
 });
 
@@ -166,7 +166,7 @@ app.get('/cart/save',authorize.authorize_user, async (req, res) => {
     res.redirect('/cart');
   } catch (err) {
     console.error('Error saving cart:', err);
-    res.render('error', { message: 'Unable to save cart. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to save cart. Please try again.' });
   }
 });
 
@@ -177,7 +177,7 @@ app.get('/cart/submit', authorize.authorize_user, async (req, res) => {
     res.render('user/cart_submit', { cart: products, user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role });
   } catch (err) {
     console.error('Error rendering cart submission page:', err);
-    res.render('error', { message: 'Unable to render cart submission page. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to render cart submission page. Please try again.' });
   }
 });
 
@@ -197,7 +197,7 @@ app.get('/products', async (req, res) => {
     res.render('products', { products, colors, user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role, url: req.url, user_role:req.signedCookies.role });
   } catch (error) {
     console.error('Error fetching products or colors:', error);
-    res.status(500).render('error', { message: 'Internal Server Error' });
+    res.status(500).render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Internal Server Error' });
   }
 });
 app.get('/search', async (req, res) => {
@@ -208,7 +208,7 @@ app.get('/search', async (req, res) => {
     res.render('search', { searchTerm, searchResults, user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role });
   } catch (error) {
     console.error('Error fetching search results:', error);
-    res.status(500).render('error', { message: 'Internal Server Error' });
+    res.status(500).render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Internal Server Error' });
   }
 });
 
@@ -223,7 +223,7 @@ app.get('/product/:id', async (req, res) => {
     res.render('product', { product: rows, user_cookie: username, url: req.url, user_role:req.signedCookies.role  });
   } catch (error) {
     console.error('Error fetching product:', error);
-    res.status(500).render('error', { message: 'Internal Server Error', user_cookie: username });
+    res.status(500).render('error', { role:req.signedCookies.role,message: 'Internal Server Error', user_cookie: username });
   }
 });
 
@@ -241,7 +241,7 @@ app.get('/order/:id', authorize.authorize_user, async (req, res) => {
     res.render('user/order', { totalPrice : price, order : order_details, products : products, user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role});
   } catch (error) {
     console.error('Error fetching order details:', error);
-    res.status(500).render('error', { message: 'Internal Server Error' });
+    res.status(500).render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Internal Server Error' });
   }
 });
 
@@ -260,7 +260,7 @@ app.get('/order_confirm', authorize.authorize_user, async (req, res) => {
     res.redirect('/order/' + order_id)
   } catch (err) {
     console.error('Error processing order:', err);
-    res.render('error', { message: 'Unable to process the order. Please try again.' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Unable to process the order. Please try again.' });
   }
 });
 
@@ -286,7 +286,7 @@ app.post('/add_product', authorize.authorize_admin, async (req, res) => {
       res.redirect('/products');
     } catch (err) {
       console.error('Error adding product:', err);
-      res.render('error', { message: 'Error adding product' });
+      res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Error adding product' });
     }
 });
 
@@ -297,7 +297,7 @@ app.get('/admin/modify/:id',authorize.authorize_admin, async (req, res) => {
     res.render('admin/modify', { user_cookie: req.signedCookies.user, role:req.signedCookies.role, product: product});
   } catch (err) {
   console.error('Error modifying product:', err);
-  res.render('error', { message: 'Error modifying product' });
+  res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Error modifying product' });
 }
 });
 
@@ -308,7 +308,7 @@ app.get('/admin/delete/:id',authorize.authorize_admin, async (req, res) => {
     res.render('admin/modify', { user_cookie: req.signedCookies.user, role:req.signedCookies.role, users: users });
   } catch (err) {
     console.error('Error deleting product:', err);
-    res.render('error', { message: 'Error deleting product' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Error deleting product' });
   }
 });
 
@@ -318,7 +318,7 @@ app.get('/users',authorize.authorize_admin, async (req, res) => {
     res.render('admin/users', { user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role, users: users });
   } catch (err) {
     console.error('Error showing users:', err);
-    res.render('error', { message: 'Error showing users' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Error showing users' });
   }
 });
 
@@ -331,7 +331,7 @@ app.get('/admin/orders',authorize.authorize_admin, async (req, res) => {
     res.render('admin/orders', { user_cookie: req.signedCookies.user, role:req.signedCookies.role, role:req.signedCookies.role, orders : orders });
   } catch (err) {
     console.error('Error showing orders:', err);
-    res.render('error', { message: 'Error showing orders' });
+    res.render('error', { user_cookie:req.signedCookies.user,role:req.signedCookies.role, message: 'Error showing orders' });
   }
 });
